rename benchmark function to multivariate and combine utils into benchmarking module
